deploy/pyproject.toml
[project]
name = "deploy-opencode"
version = "0.1.0"
description = "Convert Claude Code plugins to OpenCode format"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
deploy-opencode = "deploy_opencode.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/deploy_opencode"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"

[dependency-groups]
dev = [
    "pytest>=8.0",
]
